uncomment code
this code was comment because i thought it changes nothing but in time i noticed that the main character can go between two ladders, separated with void block, and no fall down. This piece of code correcting it. Also i added other keys for playing like WASD, F2 - for restarting the game, LShift and Enter to crash the block